requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.5.0  # https://github.com/evansd/whitenoise
redis==4.6.0  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==4.2.2  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.1.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
#drf-spectacular==0.26.3  # https://github.com/tfranzel/drf-spectacular
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
